Use Terraform's 'etag' attribute for change detection
In an earlier change we started using the file contents as part of the
resource identifier as a hack to make Terraform detect file changes.
Terraform now has first-class support for this via the etag attribute, so
we'll just use that and go back to using the relative path as the resource
identifier.

This means that Terraform should see changes to files in-place as changes,
rather than as the deletion of one resource and the creation of another.